Add files via upload
Added a folder that contains the figures from the paper as .png files. 
Added a folder that contains the data needed for the figure builders for figures 2, 3, 4, and 9. The data generation for this figures is quite long, it has been added hear so the data generators do not need to be run to view the results.